requirements.txt
vietocr~=0.3.2
Pillow~=8.0.1
